# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmPay API",
        "description": "API pour l'application OmPay - Système de paiement mobile sénégalais avec authentification double niveau",
        "contact": {
            "email": "contact@ompay.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        },
        {
            "url": "https://ompay-sa.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/client/send-otp": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Envoyer un code OTP",
                "description": "Envoie un code OTP au numéro de téléphone du client pour l'authentification",
                "operationId": "sendOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone sénégalais (format: +221XXXXXXXXX)",
                                        "type": "string",
                                        "example": "+221781157773"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP envoyé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "otp": {
                                                    "description": "Code OTP (uniquement en développement)",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Le client ne possède pas de compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Numéro de téléphone invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/client/login": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Envoyer OTP pour connexion",
                "description": "Envoie un code OTP au client actif pour se connecter",
                "operationId": "clientLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone du client",
                                        "type": "string",
                                        "example": "+221781157773"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP envoyé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Compte non activé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/client/verify-otp": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Vérifier OTP et authentifier",
                "description": "Vérifie le code OTP et active le compte ou connecte le client selon le type",
                "operationId": "verifyOtpNew",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "otp"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone du client",
                                        "type": "string",
                                        "example": "+221781157773"
                                    },
                                    "otp": {
                                        "description": "Code OTP à 6 chiffres",
                                        "type": "string",
                                        "example": "482913"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "description": "Token d'accès OAuth2",
                                                    "type": "string"
                                                },
                                                "refresh_token": {
                                                    "description": "Token de rafraîchissement",
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "OTP invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/client/compte": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Récupérer le profil client",
                "description": "Récupère les informations du client connecté, son compte, ses transactions et son QR code",
                "operationId": "getClientProfile",
                "responses": {
                    "200": {
                        "description": "Profil client récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil client récupéré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "client": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "uuid-client"
                                                        },
                                                        "nom": {
                                                            "type": "string",
                                                            "example": "Diallo"
                                                        },
                                                        "prenom": {
                                                            "type": "string",
                                                            "example": "Moussa"
                                                        },
                                                        "telephone": {
                                                            "type": "string",
                                                            "example": "+221781157773"
                                                        },
                                                        "nci": {
                                                            "type": "string",
                                                            "example": "1234567890123"
                                                        },
                                                        "statut": {
                                                            "type": "string",
                                                            "example": "actif"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "compte": {
                                                    "properties": {
                                                        "numero_compte": {
                                                            "type": "string",
                                                            "example": "+221781157773"
                                                        },
                                                        "solde": {
                                                            "type": "integer",
                                                            "example": 12500
                                                        },
                                                        "statut": {
                                                            "type": "string",
                                                            "example": "actif"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "transactions": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "type": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "transfert",
                                                                    "reception",
                                                                    "retrait",
                                                                    "paiement"
                                                                ],
                                                                "example": "transfert"
                                                            },
                                                            "telephone": {
                                                                "type": "string",
                                                                "example": "+221781157774"
                                                            },
                                                            "montant": {
                                                                "type": "integer",
                                                                "example": -3000
                                                            },
                                                            "date_transaction": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-11-12T10:35:42Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "qrcode": {
                                                    "type": "string",
                                                    "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Client non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun compte trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/client/logout": {
            "post": {
                "tags": [
                    "Authentification Client"
                ],
                "summary": "Déconnexion client",
                "description": "Déconnecte le client actuellement connecté",
                "operationId": "clientLogout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Réponse d'erreur",
                "description": "Réponse en cas d'erreur",
                "properties": {
                    "success": {
                        "description": "Statut de succès",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Message d'erreur",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu via l'authentification admin ou client",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}